Moved DelegateForm to root components folder. Added Delegate page (not in navbar yet) with the DelegateForm
